implementation of 'Blueprints' and dedicated database class
