+++ pages/simulate.html
@@ -1,7 +1,7 @@
-<form action="http://127.0.0.1:2688/api/data/insert" method="post">
-    <input type="text" name="token" value="M7!ec2gx~QS§J(+Dd">
+<form action="/api/data/insert" method="get">
+    <input type="text" name="token" placeholder="insert_token">
     <input type="text" name="t" value="574" placeholder="Temperatur">
     <input type="text" name="w" value="32" placeholder="Gewicht">
-    <input type="text" name="b" value="723" placeholder="Luftfeuchtigkeit">
+    <input type="text" name="h" value="723" placeholder="Luftfeuchtigkeit">
     <input type="submit" value="Daten senden">
 </form>
